Refactored solution to simply delay execution
